gh-108303: Move `stmpd` to `test.support`
