+++ Lib/test/test_smtplib.py
@@ -22,9 +22,8 @@ from test.support import hashlib_helper
 from test.support import socket_helper
 from test.support import threading_helper
 from test.support import asyncore
+from test.support import smtpd
 from unittest.mock import Mock
-
-from . import smtpd
 
 
 support.requires_working_socket(module=True)
